[webapi][bluetooth] change isConnected attribute check in BluetoothDevice.html
This test case aims to verify that 'isConnected' attribute is readonly.
But the test considers that after bonding a remove device, this device
is also connected. Actually this is not obviously the case, and sometime
the device is paired but not connected, and the test is failed because of this.

I suggest that change the expected device connected 'true' value to the
real connected value and just check if it is a readonly attribute.