[webapi] Remove tct-manual-w3c-tests suite
Historically this test suite was created to save testing execution time,
by copying near all manual tests into this test suite.

However it couldn't get this target thus was not used in released package at
http://download.tizen.org/tct/2.2.1/2.2.1_r1/web-tct_2.2.1_r1.tar.gz

Because the test files are duplicated to the original test suite
and they haven't been updated for a long time; so remove it.