[tools] Apply pack scripts to cordova with higher version than 5.0.0
Before the pack scripts required the cordova version exactly 5.0.0.
If it's not 5.0.0, packing will fail. Now cordova version has already
upgrade to 5.1.1. But the scripts do not update to accommodate to this change.

On master branch this modification was done before but it's not synchronized to
Crosswalk-14 branch. So merge this change to Crosswalk-14 branch.

https://crosswalk-project.org/jira/browse/XWALK-4670