[apptools] Stop setting the Length variables in the global scope.
This partially reverts commit 501e71ca ("[apptools] Add extension file
for apptools").

Setting variables such as `x86Length` in the global scope is wrong, as
they will not be available in the test methods unless one uses a
`global` statement. If that is not done, we end up with failures like
this:

    ======================================================================
    ERROR: test_target_arch (__main__.TestCrosswalkApptoolsFunctions)
    ----------------------------------------------------------------------
    Traceback (most recent call last):
      File "crosswalk_pkg_basic.py", line 312, in test_target_arch
        x86Length = x86Length + 1
    UnboundLocalError: local variable 'x86Length' referenced before assignment